// src/generated/prisma/enums.ts
export const MembershipTier = {
  BASIC: 'BASIC',
  PROFESSIONAL: 'PROFESSIONAL',
  EXECUTIVE: 'EXECUTIVE'
} as const

export type MembershipTier = (typeof MembershipTier)[keyof typeof MembershipTier]
// ...
export const PayoutSchedule = {
  WEEKLY: 'WEEKLY',
  BIWEEKLY: 'BIWEEKLY',
  INSTANT: 'INSTANT'
} as const

export type PayoutSchedule = (typeof PayoutSchedule)[keyof typeof PayoutSchedule]